requirements.txt
replicate
streamlit
matplotlib
numpy
opencv-python-headless
pillow
toml